Fixed recent "bugix" which wasn't actually a bug.
I added a comment describing why this happens, but i'll sum it up here.
The main header needs to be of size 0 so that it will not be
overwrittten. I thought that it should *not* be of size 0 because it
actually isn't. `sizeof(memheader)` is not 0, bug it only introduces
bugs when we specify the real size. There's no danger in setting the
size to zero to my knowledge.